codex

+++ engine/plugin/plugins/ffmpeg_adapter/audio_decoder/audio_ffmpeg_decoder_plugin.cpp
@@ -346,7 +346,7 @@ Status AudioFfmpegDecoderPlugin::AssignExtraDataIfExistsLocked(const std::shared
             ctx->extradata = nullptr;
             return Status::ERROR_UNKNOWN;
         }
-        ctx->extradata_size = configSize;
+        ctx->extradata_size = static_cast<int32_t>configSize;
     }
     return Status::OK;
 }
